вынес попап из внутреннего презентера во внешний
